fix: 엔티티 생성 문제 해결
