+++ src/main/java/com/example/server/domain/entity/DiaryParticipant.java
@@ -1,8 +1,7 @@
-package com.example.server.domian.entity;
+package com.example.server.domain.entity;
 
-import com.example.server.domian.enums.ContentType;
+import com.example.server.domain.enums.ContentType;
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -12,8 +11,6 @@ import java.time.LocalDateTime;
 @Entity
 @Getter
 @NoArgsConstructor
-@AllArgsConstructor
-@Table(name = "diaryParticipant")
 public class DiaryParticipant {
 
     @Id
@@ -28,15 +25,15 @@ public class DiaryParticipant {
 
     private LocalDateTime writtenDate;
 
-    @Enumerated(EnumType.STRING)
-    private ContentType contentType;
+//    @Enumerated(EnumType.STRING)
+//    private ContentType contentType;
 
     @ManyToOne
     @JoinColumn(name = "user_id")
     private User user;
 
     @ManyToOne
-    @JoinColumn(name = "familyDiary_id")
+    @JoinColumn(name = "family_diary_id")
     private FamilyDiary familyDiary;
 }
