+++ src/main/java/com/example/server/domain/entity/Family.java
@@ -1,8 +1,6 @@
-package com.example.server.domian.entity;
+package com.example.server.domain.entity;
 
-import com.fasterxml.jackson.annotation.JsonManagedReference;
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -12,7 +10,6 @@ import java.util.List;
 @Entity
 @Getter
 @NoArgsConstructor
-@AllArgsConstructor
 public class Family {
 
     @Id
