+++ src/main/java/com/example/server/domain/entity/FamilyEvent.java
@@ -1,8 +1,7 @@
-package com.example.server.domian.entity;
+package com.example.server.domain.entity;
 
-import com.example.server.domian.enums.EventType;
+import com.example.server.domain.enums.EventType;
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -11,8 +10,6 @@ import java.time.LocalDateTime;
 @Entity
 @Getter
 @NoArgsConstructor
-@AllArgsConstructor
-@Table(name = "familyEvent")
 public class FamilyEvent {
 
     @Id
@@ -23,8 +20,8 @@ public class FamilyEvent {
 
     private String description;
 
-    @Enumerated(EnumType.STRING)
-    private EventType eventType;
+//    @Enumerated(EnumType.STRING)
+//    private EventType eventType;
 
     private LocalDateTime timeline;
 
@@ -33,7 +30,7 @@ public class FamilyEvent {
     private User user;
 
     @ManyToOne
-    @JoinColumn(name = "familyMember_id")
+    @JoinColumn(name = "family_member_id")
     private FamilyMember familyMember;
 }
