+++ src/main/java/com/example/server/domain/enums/AlarmType.java
@@ -1,0 +1,5 @@
+package com.example.server.domain.enums;
+
+public enum AlarmType {
+
+}
