+++ src/main/java/com/example/server/domain/enums/ContentType.java
@@ -1,0 +1,4 @@
+package com.example.server.domain.enums;
+
+public enum ContentType {
+}
