+++ src/main/java/com/example/server/domain/enums/EventType.java
@@ -1,0 +1,4 @@
+package com.example.server.domain.enums;
+
+public enum EventType {
+}
