+++ src/main/java/com/example/server/domain/enums/RuleType.java
@@ -1,0 +1,4 @@
+package com.example.server.domain.enums;
+
+public enum RuleType {
+}
