+++ src/main/java/com/example/server/domain/enums/Status.java
@@ -1,4 +1,4 @@
-package com.example.server.domian.enums;
+package com.example.server.domain.enums;
 
 public enum Status {
     ACTIVE, INACTIVE
