--- src/main/java/com/example/server/domian/entity/Alarm.java
@@ -1,34 +1,0 @@
-package com.example.server.domian.entity;
-
-import com.example.server.domian.enums.AlarmType;
-import com.example.server.domian.enums.From;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-
-@Entity
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class Alarm {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Enumerated(EnumType.STRING)
-    private AlarmType alarmType;
-
-    @Enumerated(EnumType.STRING)
-    private From from;
-
-    @ManyToOne
-    @JoinColumn(name = "user_id")
-    private User user;
-}
-
-
-
-
