--- src/main/java/com/example/server/domian/enums/AlarmType.java
@@ -1,5 +1,0 @@
-package com.example.server.domian.enums;
-
-public enum AlarmType {
-
-}
