--- src/main/java/com/example/server/domian/enums/AnniversaryType.java
@@ -1,4 +1,0 @@
-package com.example.server.domian.enums;
-
-public enum AnniversaryType {
-}
