--- src/main/java/com/example/server/domian/enums/ContentType.java
@@ -1,4 +1,0 @@
-package com.example.server.domian.enums;
-
-public enum ContentType {
-}
