--- src/main/java/com/example/server/domian/enums/From.java
@@ -1,5 +1,0 @@
-package com.example.server.domian.enums;
-
-public enum From {
-
-}
