--- src/main/java/com/example/server/domian/enums/MoodType.java
@@ -1,4 +1,0 @@
-package com.example.server.domian.enums;
-
-public enum MoodType {
-}
