+++ src/main/java/com/example/server/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.example.server.repository;
 
-import com.example.server.domian.entity.User;
+import com.example.server.domain.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
